+++ CHANGELOG.md
@@ -5,7 +5,11 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 
 ## Unreleased
 
+### Changed
+- Mesos check supports multiple servers.
+
 ## [0.0.2] - 2015-07-14
+
 ### Changed
 - updated sensu-plugin gem to 1.2.0
 
@@ -16,4 +20,3 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 
 ### Added
 - initial release
-